+++ iroh-relay/src/server.rs
@@ -1156,17 +1156,13 @@ impl hyper::service::Service<Request<Incoming>> for CaptivePortalService {
 
 #[cfg(test)]
 mod tests {
-    use std::{
-        net::{Ipv4Addr, Ipv6Addr},
-        sync::Arc,
-        time::Duration,
-    };
+    use std::{net::Ipv4Addr, sync::Arc, time::Duration};
 
     use http::StatusCode;
     use iroh_base::{EndpointId, RelayUrl, SecretKey};
-    use iroh_dns::dns::{BoxIter, DnsError, DnsResolver, Resolver, TxtRecordData};
+    use iroh_dns::dns::DnsResolver;
     use n0_error::{Result, StackResultExt, StdResultExt};
-    use n0_future::{SinkExt, StreamExt, boxed::BoxFuture};
+    use n0_future::{SinkExt, StreamExt};
     use n0_tracing_test::traced_test;
     use rand::{RngExt, SeedableRng};
     use tracing::{info, instrument};
@@ -1182,6 +1178,7 @@ mod tests {
             handshake,
             relay::{ClientToRelayMsg, Datagrams, RelayToClientMsg},
         },
+        test_utils::StaticResolver,
         tls::{self, CaRootsConfig, default_provider},
     };
 
@@ -1575,38 +1572,8 @@ mod tests {
         Ok(())
     }
 
-    /// A resolver that hands out fixed IPv4 and IPv6 addresses for every host.
-    #[derive(Debug, Clone)]
-    struct StaticResolver {
-        v4: Vec<Ipv4Addr>,
-        v6: Vec<Ipv6Addr>,
-    }
-
-    impl Resolver for StaticResolver {
-        fn lookup_ipv4(&self, _host: String) -> BoxFuture<Result<BoxIter<Ipv4Addr>, DnsError>> {
-            let addrs: BoxIter<_> = Box::new(self.v4.clone().into_iter());
-            Box::pin(std::future::ready(Ok(addrs)))
-        }
-
-        fn lookup_ipv6(&self, _host: String) -> BoxFuture<Result<BoxIter<Ipv6Addr>, DnsError>> {
-            let addrs: BoxIter<_> = Box::new(self.v6.clone().into_iter());
-            Box::pin(std::future::ready(Ok(addrs)))
-        }
-
-        fn lookup_txt(&self, _host: String) -> BoxFuture<Result<BoxIter<TxtRecordData>, DnsError>> {
-            let records: BoxIter<_> = Box::new(std::iter::empty());
-            Box::pin(std::future::ready(Ok(records)))
-        }
-
-        fn clear_cache(&self) {}
-
-        fn reset(&self) -> Box<dyn Resolver> {
-            Box::new(self.clone())
-        }
-    }
-
-    /// A relay client that prefers IPv6 falls back to IPv4 when the advertised IPv6
-    /// address is unreachable.
+    /// Regression test: A relay client that prefers IPv6 falls back to IPv4
+    /// when the advertised IPv6 address is unreachable.
     #[tokio::test]
     #[traced_test]
     async fn test_relay_client_falls_back_to_ipv4() -> Result {
@@ -1618,21 +1585,18 @@ mod tests {
         let server = Server::spawn(config).await?;
         let addr = server.http_addr().expect("http relay address");
 
-        // `relay.test` resolves to the relay's real IPv4 address alongside an
-        // unreachable IPv6 address from the RFC 3849 documentation prefix, standing
-        // in for the stale AAAA record from the issue.
-        let resolver = DnsResolver::custom(StaticResolver {
-            v4: vec![Ipv4Addr::LOCALHOST],
-            v6: vec!["2001:db8::dead".parse().expect("valid IPv6")],
-        });
+        // Resolves to both the real IPv4 address and an unreachable IPv6 address.
+        let resolver = StaticResolver::new(
+            vec![Ipv4Addr::LOCALHOST],
+            vec!["2001:db8::dead".parse().expect("valid IPv6")],
+        );
         let url: Url = format!("http://relay.test:{}", addr.port())
             .parse()
             .expect("valid relay url");
 
-        // Force the IPv6 preference, as a client with working IPv6 connectivity
-        // would have after a net report.
         let client = ClientBuilder::new(url, SecretKey::generate(), resolver)
             .tls_client_config(tls::make_dangerous_client_config())
+            // Force IPv6 preference
             .address_family_selector(|| true);
 
         tokio::time::timeout(Duration::from_secs(10), client.connect())
